backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
redis==5.0.1
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
pytest==8.0.0
langchain==0.1.0
openai==1.12.0
pydantic==2.6.0
pydantic-settings==2.2.1
pinecone-client==3.0.2
azure-storage-file-datalake==12.14.0
azure-eventgrid==4.16.0
numpy==1.26.3